Update dmg build script; misc tweaks

+++ builds/macx/CreateDmg.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # Trevor SANDY
-# Last Update 14 February 2017
+# Last Update 28 August 2017
 # To run:
 # $ chmod 755 CreateDmg.sh
 # $ ./CreateDmg.sh
@@ -9,12 +9,32 @@ ME="$(basename "$(test -L "$0" && readlink "$0" || echo "$0")")"
 CWD=`pwd`
 BUILD_DATE=`date "+%Y%m%d"`
 # logging stuff
-LOG="${CWD}/$ME.log"
+# increment log file name
+f="${CWD}/$ME"
+ext=".log"
+if [[ -e "$f$ext" ]] ; then
+    i=1
+    f="${f%.*}";
+    while [[ -e "${f}_${i}${ext}" ]]; do
+        let i++
+    done
+    f="${f}_${i}${ext}"
+else
+   f="${f}${ext}"
+fi
+# output log file
+LOG="$f"
 exec > >(tee -a ${LOG} )
 exec 2> >(tee -a ${LOG} >&2)
 
 echo "Start $ME..."
 export PATH=~/Qt/IDE/5.7/clang_64:~/Qt/IDE/5.7/clang_64/bin:$PATH
+
+echo
+echo "Enter n to download source or any key "
+echo "to use existing source if available."
+read -n 1 -p "Do you want to continue with these options? : " input
+echo
 
 echo "1. create DMG build working directory"
 if [ ! -d dmgbuild ]
@@ -23,15 +43,29 @@ then
 fi
 cd dmgbuild
 
-echo "2. download source"
-if [ -d lpub3d ]
+if [ ! -d lpub3d_macos_3rdparty ]
 then
-    rm -rf lpub3d
+    echo "2a. download lpub3d 3rd party renderers"
+    git clone https://github.com/trevorsandy/lpub3d_macos_3rdparty.git
+  else
+    echo "2a. renderer folder exist. skipping download"
 fi
-git clone https://github.com/trevorsandy/lpub3d.git
+
+if [ -d lpub3d ] && [ "$input" = "n" ] || [ "$input" = "N" ]
+then
+    echo "2b. download lpub3d source"
+    rm -rf lpub3d
+    git clone https://github.com/trevorsandy/lpub3d.git
+  else
+    echo "2b. lpubsource exist skipping download"
+fi
+
+echo "2c. checkout integrate-renderers"
+cd lpub3d
+## TEMPERARY ##
+#git checkout integrate-renderers
 
 echo "3. capture version and date info"
-cd lpub3d
 #         1 2  3  4   5       6    7  8  9       10
 # format "2 0 20 17 663 410fdd7 2017 02 12 19:50:21"
 FILE="builds/utilities/version.info"
@@ -47,7 +81,7 @@ read VER_MAJOR VER_MINOR VER_PATCH VER_REVISION  VER_BUILD VER_SHA_HASH <<< ${VE
 VERSION=${VER_MAJOR}"."${VER_MINOR}"."${VER_PATCH}
 APP_VERSION=${VERSION}"."${VER_BUILD}
 APP_VERSION_LONG=${VERSION}"."${VER_REVISION}"."${VER_BUILD}_${BUILD_DATE}
-echo "WORK_DIR..........${WORK_DIR}"
+#echo "WORK_DIR..........${WORK_DIR}"
 echo "VER_MAJOR.........${VER_MAJOR}"
 echo "VER_MINOR.........${VER_MINOR}"
 echo "VER_PATCH.........${VER_PATCH}"
@@ -59,9 +93,16 @@ echo "APP_VERSION.......${APP_VERSION}"
 echo "APP_VERSION_LONG..${APP_VERSION_LONG}"
 echo "BUILD_DATE........${BUILD_DATE}"
 
-echo "4. get ldraw archive libraries"
-curl "http://www.ldraw.org/library/updates/complete.zip" -o "mainApp/extras/complete.zip"
-curl "http://www.ldraw.org/library/unofficial/ldrawunf.zip" -o "mainApp/extras/lpub3dldrawunf.zip"
+if [ ! -f "mainApp/extras/complete.zip" ]
+then
+  echo "4a. get ldraw official library archive"
+  curl "http://www.ldraw.org/library/updates/complete.zip" -o "mainApp/extras/complete.zip"
+fi
+if [ ! -f "mainApp/extras/lpub3dldrawunf.zip" ]
+then
+  echo "4b. get ldraw unofficial library archive"
+  curl "http://www.ldraw.org/library/unofficial/ldrawunf.zip" -o "mainApp/extras/lpub3dldrawunf.zip"
+fi
 
 echo "5. configure source"
 qmake LPub3D.pro -spec macx-clang CONFIG+=x86_64 && /usr/bin/make qmake_all
@@ -121,7 +162,7 @@ EOF
 
 echo "13. create dmg packages"
 DMGDIR=../../../DMGS
-if [ -d ${DMGDIR} ] 
+if [ -d ${DMGDIR} ]
 then
     rm -R ${DMGDIR}
 fi
@@ -138,4 +179,4 @@ rm -R LPub3D.app
 rm lpub3d.icns README.txt lpub3d.json lpub3dbkg.png
 
 echo "$ME Finished!"
-mv $LOG "${CWD}/dmgbuild/$ME.log"
+# mv $LOG "${CWD}/dmgbuild/$ME.log"
